Revert "Tests + Print services"
