+++ src/components/print/PrintDirective.js
@@ -2,7 +2,6 @@ goog.provide('ga_print_directive');
 
 goog.require('ga_attribution_service');
 goog.require('ga_browsersniffer_service');
-goog.require('ga_printlayer_service');
 goog.require('ga_printstyle_service');
 goog.require('ga_time_service');
 goog.require('ga_urlutils_service');
@@ -13,7 +12,6 @@ goog.require('ga_urlutils_service');
     'ga_browsersniffer_service',
     'pascalprecht.translate',
     'ga_printstyle_service',
-    'ga_printlayer_service',
     'ga_time_service',
     'ga_attribution_service',
     'ga_urlutils_service'
@@ -22,7 +20,7 @@ goog.require('ga_urlutils_service');
   module.controller('GaPrintDirectiveController', function($rootScope, $scope,
       $http, $q, $window, $translate, $timeout, gaLayers, gaMapUtils, 
       gaPermalink, gaBrowserSniffer, gaWaitCursor, gaPrintStyle,
-      gaPrintLayer, gaTime, gaAttribution, gaUrlUtils) {
+      gaTime, gaAttribution, gaUrlUtils) {
 
     var pdfLegendsToDownload = [];
     var pdfLegendString = '_big.pdf';
@@ -36,7 +34,8 @@ goog.require('ga_urlutils_service');
     var layersYears = [];
     var canceller;
     var currentMultiPrintId;
-
+    var format = new ol.format.GeoJSON();
+    var styleId = 0;
     $scope.printConfigLoaded = false;
     $scope.options.multiprint = false;
     $scope.options.movie = false;
@@ -128,6 +127,415 @@ goog.require('ga_urlutils_service');
       ctx.restore();
     };
 
+    // Encode ol.Layer to a basic js object
+    var encodeLayer = function(layer, proj) {
+      var encLayer, encLegend;
+
+      if (!(layer instanceof ol.layer.Group)) {
+        var src = layer.getSource();
+        var layerConfig = gaLayers.getLayer(layer.bodId) || {};
+        var resolution = $scope.map.getView().getResolution();
+        var minResolution = layerConfig.minResolution || 0;
+        var maxResolution = layerConfig.maxResolution || Infinity;
+
+        if (resolution <= maxResolution &&
+            resolution >= minResolution) {
+          if (src instanceof ol.source.WMTS) {
+            encLayer = $scope.encoders.layers['WMTS'].call(this, layer,
+                layerConfig);
+          } else if (src instanceof ol.source.ImageWMS ||
+              src instanceof ol.source.TileWMS) {
+            encLayer = $scope.encoders.layers['WMS'].call(this, layer,
+                layerConfig);
+          } else if (src instanceof ol.source.Vector ||
+              src instanceof ol.source.ImageVector) {
+            if (src instanceof ol.source.ImageVector) {
+              src = src.getSource();
+            }
+            encLayer = $scope.encoders.layers['Vector'].call(this, layer,
+                src.getFeatures());
+          }
+        }
+      }
+
+      if ($scope.options.legend && layerConfig.hasLegend) {
+        encLegend = $scope.encoders.legends['ga_urllegend'].call(this, layer,
+            layerConfig);
+
+        if (encLegend.classes && encLegend.classes[0] &&
+            encLegend.classes[0].icon) {
+          var legStr = encLegend.classes[0].icon;
+          if (legStr.indexOf(pdfLegendString,
+              legStr.length - pdfLegendString.length) !== -1) {
+            pdfLegendsToDownload.push(legStr);
+            encLegend = undefined;
+          }
+        }
+      }
+      return {layer: encLayer, legend: encLegend};
+    };
+
+
+    // Encoders by type of layer
+    $scope.encoders = {
+      'layers': {
+        'Layer': function(layer) {
+          var enc = {
+            layer: layer.bodId,
+            opacity: layer.getOpacity()
+          };
+          return enc;
+        },
+        'Group': function(layer, proj) {
+          var encs = [];
+          var subLayers = layer.getLayers();
+          subLayers.forEach(function(subLayer, idx, arr) {
+            if (subLayer.visible) {
+              var enc = $scope.encoders.layers['Layer'].call(this, layer);
+              var layerEnc = encodeLayer(subLayer, proj);
+              if (layerEnc && layerEnc.layer) {
+                $.extend(enc, layerEnc);
+                encs.push(enc.layer);
+              }
+            }
+          });
+          return encs;
+        },
+        'Vector': function(layer, features) {
+          var enc = $scope.encoders.layers['Layer'].call(this, layer);
+          var encStyles = {};
+          var encFeatures = [];
+          var stylesDict = {};
+
+          // Sort features by geometry type
+          var newFeatures = [];
+          var polygons = [];
+          var lines = [];
+          var points = [];
+
+          angular.forEach(features, function(feature) {
+            var geotype = feature.getGeometry().getType();
+            if (/^(Polygon|MultiPolygon|Circle|GeometryCollection)$/.
+                test(geotype)) {
+              polygons.push(feature);
+            } else if (/^(LineString|MultiLineString)$/.test(geotype)) {
+              lines.push(feature);
+            } else {
+              points.push(feature);
+            }
+          });
+          features = newFeatures.concat(polygons, lines, points);
+
+          angular.forEach(features, function(feature) {
+            var encoded = $scope.encoders.features.feature(layer, feature);
+            encFeatures = encFeatures.concat(encoded.encFeatures);
+            angular.extend(encStyles, encoded.encStyles);
+          });
+          angular.extend(enc, {
+            type: 'Vector',
+            styles: encStyles,
+            styleProperty: '_gx_style',
+            geoJson: {
+              type: 'FeatureCollection',
+              features: encFeatures
+            },
+            name: layer.bodId
+          });
+          return enc;
+        },
+        'WMS': function(layer, config) {
+          var enc = $scope.encoders.layers['Layer'].call(this, layer);
+          var source = layer.getSource();
+          var params = source.getParams();
+          var layers = params.LAYERS.split(',') || [];
+          var styles = (params.STYLES !== undefined) ?
+              params.STYLES.split(',') :
+              new Array(layers.length).join(',').split(',');
+          var url = (source.getUrls && source.getUrls()[0]) ||
+              (source.getUrl && source.getUrl());
+
+          angular.extend(enc, {
+            type: 'WMS',
+            baseURL: url.replace(/^\/\//, 'https://'),
+            layers: layers,
+            styles: styles,
+            format: 'image/' + (config.format || 'png'),
+            customParams: {
+              'EXCEPTIONS': 'XML',
+              'TRANSPARENT': 'true',
+              'CRS': 'EPSG:21781',
+              'TIME': params.TIME,
+              'MAP_RESOLUTION': getDpi($scope.layout.name, $scope.dpi)
+            },
+            singleTile: config.singleTile || false
+          });
+          return enc;
+        },
+
+        'matrixIds': function(tilegrid, extent) {
+          var matrixIds = [];
+          var ids = tilegrid.getMatrixIds();
+          var resolutions = tilegrid.getResolutions();
+          var defaultExtent = gaMapUtils.defaultExtent;
+          angular.forEach(resolutions, function(value, key) {
+            var resolution = parseFloat(value);
+            var z = tilegrid.getZForResolution(resolution);
+            var tileSize = tilegrid.getTileSize(z);
+            var topLeftCorner = tilegrid.getOrigin(z);
+            var minX = topLeftCorner[0];
+            var maxY = topLeftCorner[1];
+            var maxX = extent[2] || defaultExtent[2];
+            var minY = extent[1] || defaultExtent[1];
+            var topLeftTile = tilegrid
+                .getTileCoordForCoordAndZ([minX, maxY], z);
+            var bottomRightTile = tilegrid
+                .getTileCoordForCoordAndZ([maxX, minY], z);
+            var tileWidth = 1 + bottomRightTile[1] - topLeftTile[1];
+            var tileHeight = 1 + topLeftTile[2] - bottomRightTile[2];
+            var matrix = {
+              identifier: ids[key],
+              resolution: resolution,
+              topLeftCorner: tilegrid.getOrigin(z),
+              tileSize: [tileSize, tileSize],
+              matrixSize: [tileWidth, tileHeight]
+            };
+            this.push(matrix);
+
+          }, matrixIds);
+
+          return matrixIds;
+        },
+        // Dimensions
+        'dimensions': function(dimensions) {
+          var params = {};
+
+          angular.forEach(dimensions, function(value, key) {
+            params[key.toUpperCase()] = value;
+          });
+
+          return params;
+        },
+
+        'WMTS': function(layer, config) {
+
+          // config is not defined for external WMTS
+          // For internal WMTS layer, we use the simplified
+          // mapfish print protocol, and the standard for
+          // external WMTS layers.
+          // See http://www.mapfish.org/doc/print/protocol.html#wmts
+
+          var isExternalWmts = angular.equals(config, {});
+
+          var enc = $scope.encoders.layers['Layer'].call(this, layer);
+          var source = layer.getSource();
+          var tileGrid = source.getTileGrid();
+          var extent = layer.getExtent();
+
+          var requestEncoding = source.getRequestEncoding() || 'REST';
+
+          if (!config.background && layer.visible && config.timeEnabled) {
+            if (!layer.time) {
+              return;
+            }
+            layersYears.push(layer.time);
+          }
+
+          // resourceURL for RESTful, service endpoint for KVP
+          var url = source.getUrls()[0];
+          var baseUrl = url.replace(/^\/\//, 'https://');
+
+          if (requestEncoding == 'REST') {
+            baseUrl = baseUrl
+              .replace(/\{Time\}/i, '{TIME}')
+              .replace(/\{/g, '%7B')
+              .replace(/\}/g, '%7D')
+              .replace(/wmts\d{1,3}\.geo\.admin\.ch/, 'wmts.geo.admin.ch');
+          }
+
+          var wmts_dimensions = $scope.encoders.layers['dimensions']
+              .call(this, source.getDimensions());
+          // common config
+          angular.extend(enc, {
+            type: 'WMTS',
+            layer: source.getLayer(),
+            version: source.getVersion() || '1.0.0',
+            requestEncoding: requestEncoding,
+            formatSuffix: source.getFormat().replace('image/', ''),
+            style: source.getStyle() || 'default',
+            dimensions: Object.keys(wmts_dimensions),
+            params: wmts_dimensions,
+            matrixSet: source.getMatrixSet() || '21781'
+          });
+
+        if (!isExternalWmts) {
+
+          angular.extend(enc, {
+            baseURL: baseUrl.slice(0, baseUrl.indexOf('/1.0.0')),
+            zoomOffset: tileGrid.getMinZoom(),
+            tileOrigin: tileGrid.getOrigin(),
+            tileSize: [tileGrid.getTileSize(), tileGrid.getTileSize()],
+            resolutions: tileGrid.getResolutions(),
+            maxExtent: extent
+           });
+
+        } else {
+          // use the full monty WMTS definition fo external source
+          var matrixIds = $scope.encoders.layers['matrixIds']
+              .call(this, tileGrid, extent);
+
+          angular.extend(enc, {
+            layer: source.getLayer(),
+            baseURL: baseUrl,
+            matrixIds: matrixIds
+          });
+        }
+
+        var multiPagesPrint = false;
+          if (config.timestamps) {
+            multiPagesPrint = !config.timestamps.some(function(ts) {
+              return ts == '99991231';
+          });
+        }
+        // printing time series
+        if (config.timeEnabled && gaTime.get() == undefined &&
+             multiPagesPrint) {
+           enc['timestamps'] = config.timestamps;
+        }
+
+        return enc;
+        }
+      },
+      'features': {
+        'feature': function(layer, feature, styles) {
+          var encStyles = {};
+          var encFeatures = [];
+          var encStyle = {
+            id: styleId++
+          };
+
+          // Get the styles of the feature
+          if (!styles) {
+            if (feature.getStyleFunction()) {
+              styles = feature.getStyleFunction().call(feature);
+            } else if (layer.getStyleFunction()) {
+              styles = layer.getStyleFunction()(feature);
+            } else {
+              styles = ol.style.Style.defaultFunction(feature);
+            }
+          }
+          var geometry = feature.getGeometry();
+          var styleToEncode;
+          if (styles && styles.length > 0) {
+            styleToEncode = styles[0];
+          }
+
+          // We encode the feature only if the feature has a style.
+          if (!styleToEncode) {
+            return {
+              encFeatures: [],
+              encStyles: []
+            };
+          }
+
+          // Transform an ol.geom.Circle to a ol.geom.Polygon
+          if (geometry instanceof ol.geom.Circle) {
+            geometry = gaPrintStyle.olCircleToPolygon(geometry);
+            feature = new ol.Feature(geometry);
+          }
+
+          // Handle ol.style.RegularShape by converting points to polygons
+          var image = styleToEncode.getImage();
+          if (geometry instanceof ol.geom.Point &&
+              image instanceof ol.style.RegularShape &&
+              !(image instanceof ol.style.Circle)) {
+            var scale = parseFloat($scope.scale.value);
+            var resolution = scale / UNITS_RATIO / POINTS_PER_INCH;
+            geometry = gaPrintStyle.olPointToPolygon(
+              feature.getGeometry(),
+              image.getRadius(),
+              resolution,
+              image.getPoints(),
+              image.getRotation()
+            );
+            feature = new ol.Feature(geometry);
+          }
+
+          // Encode a feature if it intersects with the extent and
+          // if the map is not rotated
+
+          if (geometry.intersectsExtent(getPrintRectangleCoords())) {
+            var encFeature = format.writeFeatureObject(feature);
+            if (!encFeature.properties) {
+              encFeature.properties = {};
+            } else {
+             // Fix circular structure to JSON
+             // see: https://github.com/geoadmin/mf-geoadmin3/issues/1213
+              delete encFeature.properties.Style;
+              delete encFeature.properties.overlays;
+            }
+            encFeature.properties._gx_style = encStyle.id;
+            encFeatures.push(encFeature);
+
+            // Encode the style of the feature added
+            angular.extend(encStyle, gaPrintStyle.olStyleToPrintLiteral(
+                styleToEncode, getDpi($scope.layout.name, $scope.dpi)));
+
+            // Apply the layer's opacity on fill and stroke
+            if (encStyle.fillOpacity) {
+              encStyle.fillOpacity *= layer.getOpacity();
+            }
+
+            if (encStyle.strokeOpacity) {
+              encStyle.strokeOpacity *= layer.getOpacity();
+            }
+
+            encStyles[encStyle.id] = encStyle;
+          }
+
+          // If a feature has a style with a geometryFunction defined, we
+          // must also display this geometry with the good style (used for
+          // azimuth).
+          for (var i = 0; i < styles.length; i++) {
+            var style = styles[i];
+            if (angular.isFunction(style.getGeometry())) {
+              var geom = style.getGeometry()(feature);
+              if (geom) {
+                var encoded = $scope.encoders.features.feature(layer,
+                     new ol.Feature(geom), [style]);
+                encFeatures = encFeatures.concat(encoded.encFeatures);
+                angular.extend(encStyles, encoded.encStyles);
+              }
+            }
+          }
+
+          return {
+            encFeatures: encFeatures,
+            encStyles: encStyles
+          };
+        }
+      },
+      'legends' : {
+        'ga_urllegend': function(layer, config) {
+          var format = '.png';
+          if ($scope.options.pdfLegendList.indexOf(layer.bodId) != -1) {
+            format = pdfLegendString;
+          }
+          var enc = $scope.encoders.legends.base.call(this, config);
+          enc.classes.push({
+            name: '',
+            icon: $scope.options.legendUrl +
+                layer.bodId + '_' + $translate.use() + format
+          });
+          return enc;
+        },
+        'base': function(config) {
+          return {
+            name: config.label,
+            classes: []
+          };
+        }
+      }
+    };
 
     var getZoomFromScale = function(scale) {
       var i, len;
@@ -215,12 +623,6 @@ goog.require('ga_urlutils_service');
       pdfLegendsToDownload = [];
       layersYears = [];
 
-      var dpi = getDpi($scope.layout.name, $scope.dpi);
-      var scaleDenom = parseFloat($scope.scale.value);
-      var printRectangeCoords = getPrintRectangleCoords();
-      var resolution = $scope.map.getView().getResolution();
-
-
       // Re order layer by z-index
       layers.sort(function(a, b) {
         return a.getZIndex() - b.getZIndex();
@@ -228,7 +630,6 @@ goog.require('ga_urlutils_service');
 
       // Transform layers to literal
       layers.forEach(function(layer) {
-
         if (!layer.visible || layer.opacity == 0) {
           return;
         }
@@ -247,41 +648,12 @@ goog.require('ga_urlutils_service');
         }
 
         // Encode layers
-        var encs, encLegend;
+        var encs;
         if (layer instanceof ol.layer.Group) {
-          encs = gaPrintLayer.encodeGroup(layer, proj, scaleDenom,
-              printRectangeCoords, resolution, dpi);
+          encs = $scope.encoders.layers['Group'].call(this,
+              layer, proj);
         } else {
-          //var enc = encodeLayer(layer, proj);
-          var layerConfig = gaLayers.getLayer(layer.bodId) || {};
-          var legendToPrint = $scope.options.legend && layerConfig.hasLegend;
-          var enc = gaPrintLayer.encodeLayer(layer, proj, scaleDenom,
-              printRectangeCoords, resolution, dpi);
-
-          if ($scope.options.legend && layerConfig.hasLegend) {
-            encLegend = gaPrintLayer.encodeLegend(layer, layerConfig,
-                $scope.options);
-
-          if (!layerConfig.background && layerConfig.visible &&
-              layerConfig.timeEnabled) {
-            if (!layer.time) {
-              return;
-            }
-            layersYears.push(layer.time);
-          }
-
-          if (encLegend.classes && encLegend.classes[0] &&
-              encLegend.classes[0].icon) {
-            var legStr = encLegend.classes[0].icon;
-            if (legStr.indexOf(pdfLegendString,
-                legStr.length - pdfLegendString.length) !== -1) {
-              pdfLegendsToDownload.push(legStr);
-              encLegend = undefined;
-              }
-            }
-          }
-          enc.legend = encLegend;
-
+          var enc = encodeLayer(layer, proj);
           if (enc && enc.layer) {
             encs = [enc.layer];
             if (enc.legend) {
@@ -322,21 +694,100 @@ goog.require('ga_urlutils_service');
 
       // Transform graticule to literal
       if ($scope.options.graticule) {
-        var graticule = gaPrintLayer.encodeGraticule(dpi);
-
+        var graticule = {
+          'baseURL': 'https://wms.geo.admin.ch/',
+          'opacity': 1,
+          'singleTile': true,
+          'type': 'WMS',
+          'layers': ['org.epsg.grid_2056'],
+          'format': 'image/png',
+          'styles': [''],
+          'customParams': {
+            'TRANSPARENT': true,
+            'MAP_RESOLUTION': getDpi($scope.layout.name, $scope.dpi)
+          }
+        };
         encLayers.push(graticule);
       }
 
       // Transform overlays to literal
       // FIXME this is a temporary solution
       var overlays = $scope.map.getOverlays();
+      var resolution = $scope.map.getView().getResolution();
 
       overlays.forEach(function(overlay) {
-        var encOverlayLayer = gaPrintLayer.encodeOverlay(overlay,
-            resolution, $scope.options);
+        var elt = overlay.getElement();
+        // We print only overlay added by the MarkerOverlayService
+        // or by crosshair permalink
+        if ($(elt).hasClass('popover')) {
+          return;
+        }
+        var center = overlay.getPosition();
+        var offset = 5 * resolution;
 
-        if (encOverlayLayer) {
-            encLayers.push(encOverlayLayer);
+        if (center) {
+          var style = 1, $elt = $(elt);
+          if ($elt.text()) {
+             style = 2;
+             if ($elt.hasClass('ga-draw-measure-tmp')) {
+               style = 3;
+             }
+          }
+          var encOverlayLayer = {
+            'type': 'Vector',
+            'styles': {
+              '1': { // Style for marker position
+                'externalGraphic':
+                  gaUrlUtils.unProxifyUrl($scope.options.markerUrl),
+                'graphicWidth': 20,
+                'graphicHeight': 30,
+                // the icon is not perfectly centered in the image
+                // these values must be the same in map.less
+                'graphicXOffset': -12,
+                'graphicYOffset': -30
+              }, '2': { // Style for measure tooltip
+                'externalGraphic':
+                  gaUrlUtils.unProxifyUrl($scope.options.bubbleUrl),
+                'graphicWidth': 97,
+                'graphicHeight': 27,
+                'graphicXOffset': -48,
+                'graphicYOffset': -27,
+                'label': $elt.text(),
+                'labelXOffset': 0,
+                'labelYOffset': 18,
+                'fontColor': '#ffffff',
+                'fontSize': 10,
+                'fontWeight': 'normal'
+              }, '3': { // Style for intermeediate measure tooltip
+                'label': $elt.text(),
+                'labelXOffset': 0,
+                'labelYOffset': 18,
+                'fontColor': '#ffffff',
+                'fontSize': 8,
+                'fontWeight': 'normal',
+                'fillColor': '#ff0000',
+                'strokeColor': '#ff0000'
+              }
+
+            },
+            'styleProperty': '_gx_style',
+            'geoJson': {
+              'type': 'FeatureCollection',
+              'features': [{
+                'type': 'Feature',
+                'properties': {
+                  '_gx_style': style
+                },
+                'geometry': {
+                  'type': 'Point',
+                  'coordinates': [center[0], center[1], 0]
+                }
+              }]
+            },
+            'name': 'drawing',
+            'opacity': 1
+          };
+          encLayers.push(encOverlayLayer);
         }
       });
 
@@ -622,6 +1073,8 @@ goog.require('ga_urlutils_service');
         gaWaitCursor.decrement();
       }
     });
+
+
   });
 
   module.directive('gaPrint',
